backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain-openai
mcp-use
